subdivide `draw` & impl `From` conversions
- new `draw` submodules `coord` & `rect`.
- add From/Into aray/tuple conversions for `Coord` & `Rect`.
- update docs.
