+++ fltk/src/draw/mod.rs
@@ -5,91 +5,12 @@ use crate::utils::FlString;
 use fltk_sys::draw::*;
 use std::ffi::{CStr, CString};
 use std::mem;
-use std::ops::{Add, Sub};
 use std::os::raw;
 
-/// Defines a rectangular bounding box
-#[derive(Copy, Clone, Debug)]
-pub struct Rectangle<T: Copy + Add<Output = T> + Sub<Output = T>> {
-    /// Leftmost corner
-    pub x: T,
-    /// Topmost corner
-    pub y: T,
-    /// Width
-    pub w: T,
-    /// Height
-    pub h: T,
-}
-
-/// `i32` Coordinates
-pub type Rect = Rectangle<i32>;
-
-/// `f64` Coordinates
-#[allow(non_camel_case_types)]
-pub type Rect_f64 = Rectangle<f64>;
-
-impl<T: Copy + Add<Output = T> + Sub<Output = T>> Rectangle<T> {
-    /// Returns a new `Rectangle` from its top-left corner position,
-    /// and the length of its sides.
-    pub fn new(x: T, y: T, width: T, height: T) -> Self {
-        Self { x, y, w: width, h: height, }
-    }
-
-    /// Returns a new `Rectangle` from the position of its `top_left`
-    /// and `bottom_right` corners.
-    pub fn from_coords(top_left: Coordinates<T>, bottom_right: Coordinates<T>) -> Self {
-        Self {
-            x: top_left.x,
-            y: top_left.y,
-            w: bottom_right.x - top_left.x,
-            h: bottom_right.y - top_left.y,
-        }
-    }
-
-    /// Returns the coordinates of the top-left corner
-    pub fn top_left(&self) -> Coordinates<T> {
-        Coordinates::new(self.x, self.y)
-    }
-
-    /// Returns the coordinates of the bottom-right corner
-    pub fn bottom_right(&self) -> Coordinates<T> {
-        Coordinates::new(self.x + self.w, self.y + self.h)
-    }
-}
-
-/// Defines a pair of `x, y` coordinates
-#[derive(Copy, Clone, Debug)]
-pub struct Coordinates<T: Copy> {
-    /// Horizontal X coordinate
-    pub x: T,
-    /// Vertical Y coordinate
-    pub y: T,
-}
-
-/// `i32` Coordinates
-pub type Coord = Coordinates<i32>;
-
-/// `f64` Coordinates
-#[allow(non_camel_case_types)]
-pub type Coord_f64 = Coordinates<f64>;
-
-impl<T: Copy> Coordinates<T> {
-    /// Returns a new pair of `x, y` coordinates
-    pub fn new(x: T, y: T) -> Self {
-        Coordinates { x, y }
-    }
-
-    /// Returns `x` interpreted as width.
-    #[inline]
-    pub fn w(&self) -> T {
-        self.x
-    }
-    /// Returns `y` interpreted as height.
-    #[inline]
-    pub fn h(&self) -> T {
-        self.y
-    }
-}
+mod coord;
+pub use coord::{Coordinates, Coord, Coord_f64};
+mod rect;
+pub use rect::{Rectangle, Rect, Rect_f64};
 
 bitflags::bitflags! {
     /// Defines the line styles supported by fltk
@@ -338,8 +259,11 @@ pub fn draw_pie(r: Rect, a: f64, b: f64) {
 }
 
 /// Sets the line style
+///
 /// # Warning
-/// You are required to change this back to [`set_line_style(LineStyle::Solid, 0)`](`crate::draw::set_line_style`) after finishing
+/// You are required to change this back to
+/// [`set_line_style(LineStyle::Solid, 0)`](crate::draw::set_line_style)
+/// after finishing
 pub fn set_line_style(style: LineStyle, width: i32) {
     unsafe {
         crate::app::open_display();
@@ -492,7 +416,7 @@ pub fn draw_yxline2(x: i32, y: i32, y1: i32, x2: i32) {
     unsafe { Fl_yxline2(x, y, y1, x2) }
 }
 
-///  Draws a vertical line from (x,y) to (x,y1) then a horizontal from (x,y1)
+/// Draws a vertical line from (x,y) to (x,y1) then a horizontal from (x,y1)
 /// to (x2,y1), then another vertical from (x2,y1) to (x2,y3)
 pub fn draw_yxline3(x: i32, y: i32, y1: i32, x2: i32, y3: i32) {
     unsafe { Fl_yxline3(x, y, y1, x2, y3) }
@@ -738,8 +662,10 @@ pub fn rtl_draw(txt: &str, x: i32, y: i32) {
 }
 
 /// Draws a series of line segments around the given box.
-/// The string must contain groups of 4 letters which specify one of 24 standard grayscale values,
-/// where 'A' is black and 'X' is white. The order of each set of 4 characters is: top, left, bottom, right.
+///
+/// The string must contain groups of 4 letters which specify one of 24 standard
+/// grayscale values, where 'A' is black and 'X' is white.
+/// The order of each set of 4 characters is: top, left, bottom, right.
 pub fn draw_frame(string: &str, x: i32, y: i32, width: i32, height: i32) {
     assert!(string.len() % 4 == 0);
     let s = CString::safe_new(string);
@@ -747,6 +673,7 @@ pub fn draw_frame(string: &str, x: i32, y: i32, width: i32, height: i32) {
 }
 
 /// Draws a series of line segments around the given box
+///
 /// Differs from frame() by the order of the line segments which is bottom, right, top, left.
 pub fn draw_frame2(string: &str, x: i32, y: i32, width: i32, height: i32) {
     assert!(string.len() % 4 == 0);
